modulize(?) cluster class
